+++ datacraft/server.py
@@ -19,11 +19,13 @@ class _Server:
     def __init__(self,
                  generator: Generator,
                  endpoint: str,
+                 port: int,
                  data_is_json: bool,
                  count_supplier: datacraft.ValueSupplierInterface,
                  delay: float = None):
         self.generator = generator
         self.endpoint = endpoint
+        self.port = port
         self.data_is_json = data_is_json
         self.count_supplier = count_supplier
         self.delay = delay
@@ -57,12 +59,15 @@ class _Server:
         """ run the Flask app """
         app = flask.Flask(__name__)
         _log.info('Adding endpoint to server: %s', self.endpoint)
-        app.add_url_rule(self.endpoint, view_func=self.callback, methods=['POST', 'GET'])
-        app.run()
+        app.add_url_rule(self.endpoint,
+                         view_func=self.callback,
+                         methods=['POST', 'GET'])
+        app.run(port=self.port)
 
 
 def run(generator: Generator,
         endpoint: str,
+        port: int,
         data_is_json: bool,
         count_supplier: datacraft.ValueSupplierInterface,
         delay: float = None):
@@ -73,9 +78,10 @@ def run(generator: Generator,
     Args:
         generator: that provides response data as dictionary
         endpoint: to serve data on i.e. /data
+        port: to serve data on e.g. 8080
         data_is_json: if the data should be returned as JSON, default is as string
         count_supplier: supplies the number of values to generate for each call
         delay: number of seconds to pause between request and response
     """
-    server = _Server(generator, endpoint, data_is_json, count_supplier, delay)
+    server = _Server(generator, endpoint, port, data_is_json, count_supplier, delay)
     server.run()