+++ tests/test_server.py
@@ -17,7 +17,7 @@ def test_server(mocker, one):
 
     gen = datacraft.parse_spec(spec).generator(1)
 
-    datacraft.server.run(gen, endpoint='/test', data_is_json=True, count_supplier=one, )
+    datacraft.server.run(gen, endpoint='/test', port=None, data_is_json=True, count_supplier=one)
 
 
 def test_server_callback_stop_iteration(mocker, one):
@@ -26,7 +26,7 @@ def test_server_callback_stop_iteration(mocker, one):
 
     gen = datacraft.parse_spec(spec).generator(1)
 
-    server = datacraft.server._Server(gen, endpoint='/test', data_is_json=True, count_supplier=one)
+    server = datacraft.server._Server(gen, endpoint='/test', port=None, data_is_json=True, count_supplier=one)
 
     # two calls should trigger the StopIteration
     server.callback()
@@ -39,7 +39,7 @@ def test_server_callback(mocker, one):
     processor = datacraft.outputs.processor(format_name='json-pretty')
     gen = datacraft.parse_spec(spec).generator(1, processor=processor)
 
-    server = datacraft.server._Server(gen, endpoint='/test', data_is_json=False, count_supplier=one)
+    server = datacraft.server._Server(gen, endpoint='/test', port=None, data_is_json=False, count_supplier=one)
 
     record = server.callback()
     assert record == json.dumps({'test': 42}, indent=4)